+++ cocos/platform/android/CCApplication-android.cpp
@@ -73,7 +73,7 @@ int Application::run()
     return -1;
 }
 
-void Application::setAnimationInterval(double interval)
+void Application::setAnimationInterval(float interval)
 {
   JniMethodInfo methodInfo;
   if (! JniHelper::getStaticMethodInfo(methodInfo, "org/cocos2dx/lib/Cocos2dxRenderer", "setAnimationInterval",
