+++ cocos/platform/android/java/src/org/cocos2dx/lib/Cocos2dxRenderer.java
@@ -56,7 +56,7 @@ public class Cocos2dxRenderer implements GLSurfaceView.Renderer {
     // Getter & Setter
     // ===========================================================
 
-    public static void setAnimationInterval(final double animationInterval) {
+    public static void setAnimationInterval(final float animationInterval) {
         Cocos2dxRenderer.sAnimationInterval = (long) (animationInterval * Cocos2dxRenderer.NANOSECONDSPERSECOND);
     }
